+++ src/main/java/chess/domain/ChessRoomRepository.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.domain;
 
 import chess.domain.Room;
