+++ src/main/java/chess/repository/ChessGameRepositoryImpl.java
@@ -2,6 +2,7 @@ package chess.repository;
 
 import chess.dao.GameDao;
 import chess.domain.ChessGame;
+import chess.domain.ChessGameRepository;
 import chess.domain.team.BlackTeam;
 import chess.domain.team.WhiteTeam;
 import dto.MoveDto;
