+++ src/main/java/chess/repository/TeamRepositoryImpl.java
@@ -3,6 +3,7 @@ package chess.repository;
 import chess.dao.PieceDao;
 import chess.dao.TeamDao;
 import chess.domain.Position;
+import chess.domain.TeamRepository;
 import chess.domain.piece.Piece;
 import chess.domain.team.BlackTeam;
 import chess.domain.team.Team;
